separate windows and unix for block devices
