+++ diskfs_windows.go
@@ -1,0 +1,10 @@
+package diskfs
+
+import (
+	"errors"
+)
+
+// getSectorSizes get the logical and physical sector sizes for a block device
+func getSectorSizes(f *os.File) (int64, int64, error) {
+	return 0, 0, errors.New("block devices not supported on windows")
+}
